Remove env_set function from callbacks module
This wrapper was intended to improve safety by copying the data provided
to it (since the libretro environment callback takes a *mut and can do
whatever it wants with it) however really it was a falsely reassuring
situation, and just complicates trying to pass e.g. arrays. This is why
it was removed.

Now, the specific environment setting functions in callbacks (e.g.
env_set_pixel_format will be fully responsible for upholding the
invariants that env_raw requires.